+++ tests/data/test_update_db_status.py
@@ -12,7 +12,8 @@ import unittest
 from click.testing import CliRunner
 
 from src import __version__
-from src.features.smarterdb import SmarterInfo, SampleSheep, CountrySheep
+from src.features.smarterdb import (
+    SmarterInfo, SampleSheep, Country, SampleGoat)
 from src.data.common import WORKING_ASSEMBLIES, PLINK_SPECIES_OPT
 from src.data.update_db_status import main as update_db_status
 
@@ -28,7 +29,7 @@ class UpdateDBStatusTest(
         self.plink_specie_opt = json.loads(json.dumps(PLINK_SPECIES_OPT))
 
         # create a SampleSheep object with a country
-        self.sample = SampleSheep(
+        sample = SampleSheep(
             original_id="test-1",
             smarter_id="ITOA-TEX-000000001",
             country="Italy",
@@ -40,9 +41,46 @@ class UpdateDBStatusTest(
             chip_name=self.chip_name,
             alias="alias-1",
         )
-        self.sample.save()
+        sample.save()
+        self.samples = [sample]
+
+        # add two more goat samples
+        sample = SampleGoat(
+            original_id="test-1",
+            smarter_id="ITCH-MER-000000001",
+            country="Italy",
+            species="Goat",
+            breed="Merino",
+            breed_code="Mer",
+            dataset=self.dataset,
+            type_="background",
+            chip_name=self.chip_name,
+            alias="alias-1",
+        )
+        sample.save()
+        self.samples.append(sample)
+
+        sample = SampleGoat(
+            original_id="test-2",
+            smarter_id="FRCH-MER-000000002",
+            country="France",
+            species="Goat",
+            breed="Merino",
+            breed_code="Mer",
+            dataset=self.dataset,
+            type_="background",
+            chip_name=self.chip_name,
+            alias="alias-1",
+        )
+        sample.save()
+        self.samples.append(sample)
 
         self.runner = CliRunner()
+
+    def tearDown(self):
+        SampleSheep.objects.delete()
+        SampleGoat.objects.delete()
+        Country.objects.delete()
 
     def test_help(self):
         result = self.runner.invoke(update_db_status, ["--help"])
@@ -83,10 +121,16 @@ class UpdateDBStatusTest(
         result = self.runner.invoke(update_db_status, [])
 
         self.assertEqual(0, result.exit_code)
-        qs = CountrySheep.objects.all()
+        qs = Country.objects.all()
 
-        self.assertEqual(qs.count(), 1)
+        self.assertEqual(qs.count(), 2)
 
-        italy = qs.get()
-        self.assertEqual(italy.name, "Italy")
+        italy = Country.objects.get(name="Italy")
+
         self.assertEqual(italy.alpha_2, "IT")
+        self.assertListEqual(italy.species, ["Sheep", "Goat"])
+
+        france = Country.objects.get(name="France")
+
+        self.assertEqual(france.alpha_2, "FR")
+        self.assertListEqual(france.species, ["Goat"])