+++ tests/features/test_smarterdb.py
@@ -14,7 +14,7 @@ from unittest.mock import patch
 from src.features.smarterdb import (
     VariantSheep, Location, SampleSheep,
     SmarterDBException, getSmarterId, Breed, get_or_create_breed, Dataset,
-    BreedAlias, get_or_create_sample, SEX, get_sample_type, CountrySheep)
+    BreedAlias, get_or_create_sample, SEX, get_sample_type, Country)
 
 from ..common import MongoMockMixin, SmarterIDMixin
 
@@ -24,17 +24,22 @@ DATA_DIR = pathlib.Path(__file__).parents[1] / "fixtures"
 
 class CountryTestCase(MongoMockMixin, unittest.TestCase):
     def setUp(self):
-        self.country = CountrySheep(name="Italy")
+        self.country = Country(name="Italy", species="Sheep")
+        self.country.save()
+
+    def tearDown(self):
+        Country.objects.delete()
 
     def test_init(self):
         self.assertEqual(self.country.alpha_2, "IT")
         self.assertEqual(self.country.name, "Italy")
+        self.assertListEqual(self.country.species, ["Sheep"])
 
     def test_str(self):
         self.assertEqual(str(self.country), "Italy (IT)")
 
     def test_no_official_name(self):
-        country = CountrySheep(name="Barbados")
+        country = Country(name="Barbados")
         self.assertEqual(str(country), "Barbados (BB)")
         self.assertIsNone(country.official_name)